requirements.txt
flask
numpy
Pillow
gunicorn
